Pi Game
For this project, I wanted to make a fun and innovative game centered around the idea of 'pie', by using three different forms of it: the microcontroller Raspberry Pi, the number pi, as well pie. The idea of the game is that users have to insert the correct digits of pi in order, up to the 15th digit. If they arrive there successfully, they are congratulated by a LED lights display, and are spared of a face full of whipped cream. But if, at any point, the user enters an incorrect number, then a servo triggers the game to pie the user in the face.